Change PrefetchHelper to template class
Refactor PrefetchHelper, QueryMetadataJob, DownloadItemsJob
to work with different roots types.

Relates-To: OLPEDGE-1900
